requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.8
h11==0.14.0
idna==3.10
joblib==1.4.2
nltk==3.9.1
numpy==2.2.2
pandas==2.2.3
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
textblob==0.19.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0

# Groq API requirements
groq==0.16.0
supabase